Brett_6/10/25 progress
Progress on my chicken wing radius and ulna. I finished the rough segmentation of the radius using paint tool, eraser, and fill between slices, and I'm almost done with the ulna, using the same tools. Currently at the point where I can fill between slices, but open checking it with the ulna segmentation it applies weird, out-of-place segments for the radius bone.